range-based forへ書き換え

+++ TankFlanker/define.cpp
@@ -891,7 +891,7 @@ void UIS::debug(float fps, float time) {
 	DrawFormatString(100, 100 + 0, c_ffffff, "%05.2ffps (%.2fms)", fps, time);
 	DrawFormatString(100, 100 + 18, c_ffffff, "%d(%.2fms)", 0, waydeb[0]);
 	for (size_t j = 1; j < std::size(waydeb); ++j) {
-		DrawFormatString(100, 100 + 18 + j * 18, c_ffffff, "%d(%.2fms)", j, waydeb[j] - waydeb[j - 1]);
+		DrawFormatString(100, int(100 + 18 + j * 18), c_ffffff, "%d(%.2fms)", j, waydeb[j] - waydeb[j - 1u]);
 	}
 }
 //
